More obvious "add a subpage" when there isn't one
It is not obvious you can click the italic grey text to add a subpage (when there isn't one). In fact i only found out you could do it reading the code.

I added the "normal" button to invoke the model window to make it more obvious.
I removed the action on the span element.