Optional widgets
A widget can decide based on any factor that I should not be displayed in the
panel (e.g. a version-widget can hide itself from any user who is not an admin
as the presented information would not be very useful).

Instead of returning and `array('title', 'html')` the widget code returns false.
In this case the dashboard does not create a new widget box.