image render size in file list variable
Most hi-res stock photo's and images directly from camera's are larger than the 2000 px max resolution that was set for rendering the image in the thumbnail. By making this variable there's a choice to go for some additional load, but happy end users because they can see their images in the thumbnails. :-)