Time field - default now included, fix #449
Includes a 'default: now' option for the time field and also fixes the calculation of the nearest time intervall to (pre-)select the correct select option. Thus, the default option does can be set to a time that does not match the specified intervall, but the panel will still select the nearest available time intervall option now.