Fixes tags autocomplete in structure fields, #327
This commit fixes the autocomplete for tags fields within structure field forms.

Frankly, this does not seem to be the nicest solution. It might be better to find a more radical solution to better deal with yaml fields - so that `pluck()` could actually also work with structure subfields.

Especially the loop added to `app/controllers/api/helpers.php` might rather belong somewhere in the Collection class or so.

But as I got this solution working for now, I thought I'd just share it as a pull request, which does not necessary need to be merged, but maybe helps as a starting point for a discussion on this rather fundamental issue with structure (sub-)fields.