Textarea field extends Textfield instead BaseField
As came up in this discussion (https://github.com/getkirby/panel/pull/596), it seems to be more logical that the TextareaField extends the TextField not just the BaseField.

With that change, changes to the TextField as suggested by the feature PR would refer much more logically to all text field types.