Add proper sorting for queries
Currently the options of 'pages' and 'index' queries are
automatically sorted by title. This is not always the desired
behaviour and has been removed.

Added two new query options: 'sort' and 'order'. Sort specifies
the field to sort by, without curly braces. Default is false, so
in most cases the order will not change.

To replicate the current default behaviour for pages and index,
sort has to be set to 'title'.