Apply the number field validation even if min=0
The value of `$this->min` is `0` by default, which is falsy. That's why that validation actually never applied.

WARNING: This is a breaking change because until now negative values worked even though the default minimum is set to `0`. This value will need to be manually set to `false` in the blueprint to allow negative values in the future. It was a bug, but this may lead to another one. ;)
Maybe the default value should be changed to `false` anyway.