Fieldoptions from Field
Adds the possibility to get fieldoptions from another field (e.g. a tag field) so that options can be defined dynamically.

```
favoritedrink:
    label: Your favorite drink
    type:  select
    options: field
    field:
      page:  ../
      name: availabledrinks
      seperator: ;
```

`page` enables to references a relative path to the page of the source field (parent page in the example). `name` is the name of the source field (`tags` by default). `seperator` lets you define the seperator the field value will be split by (default is `,`).