Fixes tags autocomplete for files, fix #699
Fixes/adds autocomplete for tags fields of files (by also passing the context - page/file - of the field to the autocomplete method).

It also introduces a new index option: `index: self` which only refers to the page of the field itself. This can be handy in several scenarios:
- Having two tags field on the same page, the second should be completed with values from the first but only from the same page (e.g. "categories" and "categories to highlight")
- Having a tags field in a structure field that should only be autocompleted with values from other structure entries of the same page.
- Having a file field that should be autocompleted with values from other files of only the same page.

To have tags field autocomplete working inside of a structure field this PR is crucial: #751