Fieldoptions: better relative paths resolving
Enables relative paths up to $site and back down again (e.g. `../../projects/project-a`).

Sure, you could just start from the toplevel with `projects/project-a`. But given the requests I have received just weekend in relation to the new "fieldoptions from tags field", quite a few people seem to prefer to do it this way. Certainly, not the most elegant way, but implementing this fix would provide consistency and ensure that it's still working no matter which way you try.