Fix #447: Proper support for custom folder setup
Previously the Panel would break if `kirby()->urls()->index()` was called (not set!) inside the `site.php` file for custom folder setups. The reason behind this is that the `index()` method caches its result and then the Panel's `index.php` would think that the URL has been set manually and not apply the URL fix.

This commit together with getkirby/kirby#477 fixes this behavior by checking instead if the URL has been detected or really set manually. This is a more reliable approach here.